pyproject.toml
[project]
name = "leadflow-ai"
version = "0.1.0"
description = "AI-powered lead automation suite"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["leads", "automation", "ai", "claude", "pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "gspread>=6.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "requests>=2.31",
    "rich>=13.0",
]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
